board/raspberrypi: RAUC: update firmware slots
The RAUC bundle updates the rootfs slots; it does not update the
raspberrypi firmware slots (i.e. /boot and /boot/firmware); the
partitions are written only once at the initial copy to the SD-card.

Note: The redundant firmware slots are defined already in the RAUC
system configuration file; the custom backend uses it to determine the
boot'ed slot.

The single /boot filesystem contains the file autoboot.txt. That file is
modified at every updates by the custom backend, thus, that slot does
not need updates.

The redundant /boot/firmware filesystems contain various essencial files
for the raspberrypi firmware; the firmware blobs, its configuration file
config.txt, the kernel image, its device-tree and device-tree overlay
blobs, the kernel command-line file cmdline.txt... The filesystems do
not need updates as long as the system uses the same kernel through the
updates.

The kernel files are split in the rootfs and in the raspberrypi firmware
filesystems:
 - the kernel image and the device-tree blobs (including overlays) are
   part of the raspberrypi firmware filesystem
 - the modules are part the rootfs filesystems

Consequently, any kernel changes imply an update of both filesystems;
the rootfs for the kernel blobs on one hand, and on the other hand, the
raspberrypi firmware filesystem for the kernel modules.

This adds the support for updating the firmware slots.

The RAUC system configuration describes already the redundant firmware
slots. The bundle is renamed to buildroot since it embbeds the firmware
slots to update the two slots at once[1].

Note: The non-redundant slot for the /boot filesystem is undescribed and
not updatable.

Unfortunatly, the firmware filesystem contains files related to the slot
number: the root filesytem to mount[2] stored in the file cmdline.txt.
Thus the filesystems for both firmware slots are differents and this
slightly complicates the bundle as the blob has to be slot agnostic. The
filesystem has to be modified at run-time.

Hopefully, RAUC offers bundle-based customization hooks[4] to run some
code embedded in the bundle at installation, as no handlers[5][6] were
set on the current running system (at version N) to install the upcoming
bundle (at version N+1).

The install hook creates the file cmdline.txt that contains the kernel
command-line. It includes the parameter root= with the device path of
the rootfs slot after the firmware slot is written-and-mounted by the
RAUC daemon.

The install hook[7] receives slot information from the environment[8]
such as:
 - the device path via $RAUC_SLOT_DEVICE
 - the mount point via $RAUC_SLOT_MOUNT_POINT
 - the run-time directory via $RUNTIME_DIRECTORY

For the sake of versatility and to avoid hardcoding values in the hook
(i.e. the root device path), at the cost of installation ordering.

The install hook generates the file cmdline.txt in two steps:
 - first, it creates the file cmdline.txt using the rootfs device path
   set at pre-installation of the rootfs slot
 - then, it moves the created file cmdline.txt to where it belongs at
   post-installation of the firmware slot

Important: The ordering of slot installation is crucial; the bundle
**HAS TO** install the rootfs slot first, and the firmware slot then.
The hook creates the file cmdline.txt using the root device path known
at installation of the rootfs slot only. The hook copies it then to
firmware filesystem at post-installation of the firmware slot, once it
is written-and-mounted. That ordering solution is prefered over harding
values and over parsing the file /etc/rauc/system.conf to read the
approprite (the custom backend does that already).

[1]: https://rauc.readthedocs.io/en/latest/scenarios.html#multiple-slots
[2]: https://www.kernel.org/doc/html/latest/admin-guide/kernel-parameters.html
[3]: https://www.raspberrypi.com/documentation/computers/config_txt.html#cmdline
[4]: https://rauc.readthedocs.io/en/latest/using.html#bundle-based-customization-hooks
[5]: https://rauc.readthedocs.io/en/latest/reference.html#handlers-section
[6]: https://rauc.readthedocs.io/en/latest/reference.html#sec-ref-handlers
[7]: https://rauc.readthedocs.io/en/latest/using.html#install-hooks
[8]: https://rauc.readthedocs.io/en/latest/reference.html#sec-slot-hook-interface